refactor: Move `MouseEvent` & related to new module
