+++ src/event.rs
@@ -122,6 +122,7 @@ pub(crate) mod filter;
 pub(crate) mod internal;
 pub(crate) mod key_code;
 pub(crate) mod key_modifier;
+pub(crate) mod mouse_event;
 pub(crate) mod read;
 pub(crate) mod source;
 #[cfg(feature = "event-stream")]
@@ -136,6 +137,7 @@ pub use stream::EventStream;
 
 pub use key_code::KeyCode;
 pub use key_modifier::ModifierKeyCode;
+pub use mouse_event::{MouseButton, MouseEvent, MouseEventKind};
 
 use crate::{
     csi,
@@ -677,29 +679,6 @@ impl Event {
         }
     }
 
-    /// Returns the mouse event if the event is a mouse event, otherwise `None`.
-    ///
-    /// This is a convenience method that makes code which only cares about mouse events easier to
-    /// write.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use crossterm::event;
-    ///
-    /// while let Some(mouse_event) = event::read()?.as_mouse_event() {
-    ///     // ...
-    /// }
-    /// # std::io::Result::Ok(())
-    /// ```
-    #[inline]
-    pub fn as_mouse_event(&self) -> Option<MouseEvent> {
-        match self {
-            Event::Mouse(event) => Some(*event),
-            _ => None,
-        }
-    }
-
     /// Returns the pasted string if the event is a paste event, otherwise `None`.
     ///
     /// This is a convenience method that makes code which only cares about paste events easier to write.
@@ -744,78 +723,6 @@ impl Event {
             _ => None,
         }
     }
-}
-
-/// Represents a mouse event.
-///
-/// # Platform-specific Notes
-///
-/// ## Mouse Buttons
-///
-/// Some platforms/terminals do not report mouse button for the
-/// `MouseEventKind::Up` and `MouseEventKind::Drag` events. `MouseButton::Left`
-/// is returned if we don't know which button was used.
-///
-/// ## Key Modifiers
-///
-/// Some platforms/terminals does not report all key modifiers
-/// combinations for all mouse event types. For example - macOS reports
-/// `Ctrl` + left mouse button click as a right mouse button click.
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Debug, PartialOrd, Ord, PartialEq, Eq, Clone, Copy, Hash)]
-pub struct MouseEvent {
-    /// The kind of mouse event that was caused.
-    pub kind: MouseEventKind,
-    /// The column that the event occurred on.
-    pub column: u16,
-    /// The row that the event occurred on.
-    pub row: u16,
-    /// The key modifiers active when the event occurred.
-    pub modifiers: KeyModifiers,
-}
-
-/// A mouse event kind.
-///
-/// # Platform-specific Notes
-///
-/// ## Mouse Buttons
-///
-/// Some platforms/terminals do not report mouse button for the
-/// `MouseEventKind::Up` and `MouseEventKind::Drag` events. `MouseButton::Left`
-/// is returned if we don't know which button was used.
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "derive-more", derive(IsVariant))]
-#[derive(Debug, PartialOrd, Ord, PartialEq, Eq, Clone, Copy, Hash)]
-pub enum MouseEventKind {
-    /// Pressed mouse button. Contains the button that was pressed.
-    Down(MouseButton),
-    /// Released mouse button. Contains the button that was released.
-    Up(MouseButton),
-    /// Moved the mouse cursor while pressing the contained mouse button.
-    Drag(MouseButton),
-    /// Moved the mouse cursor while not pressing a mouse button.
-    Moved,
-    /// Scrolled mouse wheel downwards (towards the user).
-    ScrollDown,
-    /// Scrolled mouse wheel upwards (away from the user).
-    ScrollUp,
-    /// Scrolled mouse wheel left (mostly on a laptop touchpad).
-    ScrollLeft,
-    /// Scrolled mouse wheel right (mostly on a laptop touchpad).
-    ScrollRight,
-}
-
-/// Represents a mouse button.
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "derive-more", derive(IsVariant))]
-#[derive(Debug, PartialOrd, Ord, PartialEq, Eq, Clone, Copy, Hash)]
-pub enum MouseButton {
-    /// Left mouse button.
-    Left,
-    /// Right mouse button.
-    Right,
-    /// Middle mouse button.
-    Middle,
 }
 
 bitflags! {
